// block-performance-1.0.0/src/library/typescript/code-file-rewrite/type.ts
// auto generated by dev/system

import { Logger } from "../../../myutils/logger.js";

export interface Input {
    filename: string;
    old_snippet: string;
    new_snippet: string;
}

export interface OutputOk {}

export interface OutputFail {}

export interface Callback<R> {
    skip: () => R;
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}

export function check_input<R>(plog: Logger, v: any, cb: { ok: () => R; fail: (err: Error) => R }): R {
    const log = plog.sub("check_input");
    log.variable("v", v);
    try {
        log.println("v must be object");
        if (typeof v !== "object" || v === null) {
            throw new Error("v is not object");
        }

        Object.keys(v).forEach((field) => {
            if (field === "filename") return;
            if (field === "old_snippet") return;
            if (field === "new_snippet") return;
            throw new Error("v contains unknown field: " + field);
        });

        log.println("v.filename must be string");
        if (typeof v.filename !== "string") {
            throw new Error("v.filename is not string");
        }

        log.println("v.old_snippet must be string");
        if (typeof v.old_snippet !== "string") {
            throw new Error("v.old_snippet is not string");
        }

        log.println("v.new_snippet must be string");
        if (typeof v.new_snippet !== "string") {
            throw new Error("v.new_snippet is not string");
        }
    } catch (err) {
        log.error(err);
        return cb.fail(err);
    }
    // nothing wrong
    log.ok();
    return cb.ok();
}

// JSON stringify value before copy to handle some specific problem
// eg. Date Object probelm
export function copy_output_ok(v: any): OutputOk {
    if (v !== undefined) {
        const v_cloned = JSON.parse(JSON.stringify(v));
        return _copy_output_ok(v_cloned);
    } else {
        return _copy_output_ok(v);
    }
}

export function _copy_output_ok(v: any): OutputOk {
    if (typeof v === "object" && v !== null) {
        const obj = {};
        return obj;
    } else {
        return {};
    }
}

// backward compatible purpose
export type Output = OutputOk;
export const copy_output = copy_output_ok;